% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes_within_1Mb_signif_SNPs.R
\name{biomart_df}
\alias{biomart_df}
\title{Dataframe-based biomart query.}
\usage{
biomart_df(dataframe, columnToFilter, mart = 38, attributes, filter)
}
\arguments{
\item{dataframe}{Dataframe with gene names.}

\item{columnToFilter}{Name of column in dataframe, which contains gene names.}

\item{mart}{Specify genome build.}

\item{attributes}{Vector of attributes to extract from BioMart.}

\item{filter}{Vector of filter to be used for BioMart query.}
}
\value{
Original dataframe together with extracted biomart query.
}
\description{
Dataframe-based biomart query.
}
