% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rs_to_loc.R
\name{convert_loc_to_rs}
\alias{convert_loc_to_rs}
\title{Convert CHR:BP locations to rs ids.}
\usage{
convert_loc_to_rs(df, dbSNP)
}
\arguments{
\item{df}{dataframe. Dataframe containing SNPs as genomic locations. Must
contain 2 columns labelled \code{CHR} and \code{BP}, with chromosome and
base pair positions, respectively. If the dataframe contains additional
columns included these will be preserved.}

\item{dbSNP}{BS genome reference snps (choose appropriate dbSNP build
dependent on genome build).}
}
\value{
Dataframe with rs ids and CHR:BP locations. **Note that some CHR:BP
  locations have more than one associated rs id, thus some filtering for
  duplicates may have to occur after conversion. We leave this to the user to
  decide how they wish to filter.** In addition, some CHR:BP locations may
  not have an associated rs id -- these will be represented by NA in the SNP
  column after conversion.
}
\description{
Convert CHR:BP locations to rs ids.
}
