% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rs_to_loc.R
\name{convert_rs_to_loc}
\alias{convert_rs_to_loc}
\title{Convert rs ids to CHR:BP locations.}
\usage{
convert_rs_to_loc(df, SNP_column, dbSNP)
}
\arguments{
\item{df}{dataframe. Dataframe containing SNPs as rs ids.}

\item{SNP_column}{chr. Name of column (in quotation marks) containing rs ids in dataframe.}

\item{dbSNP}{BS genome reference snps (choose appropriate dbSNP build
dependent on genome build)}
}
\value{
Dataframe with rs ids and CHR:BP locations.
}
\description{
Convert rs ids to CHR:BP locations.
}
