% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_GWAS_details}
\alias{get_GWAS_details}
\title{Deprecated function: tidy GWASs}
\usage{
get_GWAS_details()
}
\description{
This is only really meant to be used on the Ryten Lab server. If using
different GWASs/different servers, data tidying will have to be performed by
the user.
}
\keyword{internal}
