% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes_within_1Mb_signif_SNPs.R
\name{.query_biomart}
\alias{.query_biomart}
\title{Vector-based biomart query.}
\usage{
.query_biomart(mart = 38, attributes, filter, values)
}
\arguments{
\item{mart}{Specify genome build.}

\item{attributes}{Vector of attributes to extract from BioMart.}

\item{filter}{Vector of filter to be used for BioMart query.}

\item{values}{}
}
\value{
Original dataframe together with extracted biomart query.
}
\description{
Vector-based biomart query.
}
