% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_general_functions.R
\name{get_coloc_results}
\alias{get_coloc_results}
\title{Run coloc}
\usage{
get_coloc_results(
  df1,
  df2,
  harmonise = F,
  df1_type,
  df2_type,
  df1_beta_or_pval,
  df2_beta_or_pval,
  df1_N = NULL,
  df2_N = NULL,
  df_1_propor_cases,
  annotate_signif_SNP_df1_df2 = F,
  key_cols,
  df_1_name,
  df_2_name,
  df1_path,
  df2_path,
  p1 = 1e-04,
  p2 = 1e-04,
  p12 = 1e-05
)
}
\arguments{
\item{df1}{df. First formatted dataset for coloc analysis.}

\item{df2}{df. Second formatted dataset for coloc analysis.}

\item{harmonise}{logical. Whether you would like to modify the sign of the
beta so both datasets are with respect to the same allele}

\item{df1_type}{Either "quant" or "cc".}

\item{df2_type}{Only "quant" currently.}

\item{df1_beta_or_pval}{Either "beta" or "pval" depending on whether you want
to use betas or pvals in the coloc analysis}

\item{df2_beta_or_pval}{Either "beta" or "pval" depending on whether you want
to use betas or pvals in the coloc analysis}

\item{df1_N}{num. Number of samples in df1 (only required for quant)}

\item{df2_N}{num. Number of samples in df2 (only required for quant)}

\item{df_1_propor_cases}{num. Option to put in the proportion of cases for a
GWAS in \code{coloc::coloc.abf()}.}

\item{annotate_signif_SNP_df1_df2}{lgl. Whether to annotate results with the
signif SNPs from df1/df2 and from coloc}

\item{key_cols}{all columns needed to uniquely identify analysis - requires
appending "_1" or "_2" depending on whether col from df1 or df2
respectively}

\item{df_1_name}{chr. df1 name that will be added to results list (e.g.
"GWAS")}

\item{df_2_name}{chr. df2 name that will be added to results list (e.g.
"eQTL")}

\item{df1_path}{chr. Path to df1}

\item{df2_path}{chr. Path to df2}

\item{p1}{num. Prior probability a SNP is associated with trait 1, set to
coloc.abf default of 1e-4}

\item{p2}{num. Prior probability a SNP is associated with trait 2, set to
coloc.abf default of 1e-4}

\item{p12}{num. Prior probability a SNP is associated with both traits, set
to coloc.abf default of 1e-5}
}
\value{
List containing coloc results annotated or NULL if there are no
  overlapping SNPs or all SNPs are removed through harmonisation
}
\description{
Wrapper function that joins the two formatted datasets and performs coloc
analysis then annotates coloc results currently only allows for df1_type to
be "quant" or "cc" and df2_type to be "quant" so if running GWAS vs eQTL,
eQTL must be df2. Note: "quant" refers to a quantitative trait, while "cc"
refers to "case-control".
}
