% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_general_functions.R
\name{get_varbeta}
\alias{get_varbeta}
\title{Get varbeta}
\usage{
get_varbeta(df)
}
\arguments{
\item{df}{dataframe with the columns beta and t.stat or se}
}
\value{
df with the column varbeta added as a calculation from the beta and se/t.stat
}
\description{
Generates varbeta from the beta and the t.stat/se
}
