% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_general_functions.R
\name{join_coloc_datasets}
\alias{join_coloc_datasets}
\title{Join coloc datasets}
\usage{
join_coloc_datasets(df1, df2, harmonise = F)
}
\arguments{
\item{df1}{df. First formatted dataset for coloc analysis.}

\item{df2}{df. Second formatted dataset for coloc analysis.}

\item{harmonise}{logical. Whether you would like to modify the sign of the
beta so both datasets are with respect to the same allele}
}
\value{
Joined dataframe with data from both datasets.
}
\description{
Wrapper function that joins the two formatted datasets.
}
