% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_general_functions.R
\name{make_results_dir}
\alias{make_results_dir}
\title{Make results directory}
\usage{
make_results_dir(results_path, folder_name)
}
\arguments{
\item{results_path}{path to folder}

\item{folder_name}{name of folder to make}
}
\value{
the full path to the new folder
}
\description{
Generates output folder to store coloc results, does not overwrite the folder if it already exists
}
