% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_general_functions.R
\name{merge_coloc_summaries}
\alias{merge_coloc_summaries}
\title{Merge coloc summaries}
\usage{
merge_coloc_summaries(
  results_folder_path,
  add_signif_SNP = F,
  recursive = F,
  pattern = NULL
)
}
\arguments{
\item{results_folder_path}{chr. Path to the folder where the results are
stored}

\item{add_signif_SNP}{lgl. Determines whether to add the details of the most
signif SNPs to the summary. Default is FALSE.}

\item{recursive}{lgl. Argument used in \code{list.files}, which determines
whether to recursively look for results in every other folder within the
results_folder_path. Default is FALSE.}

\item{pattern}{lgl. Argument used in \code{list.files}, which specifies a
pattern to match for the results (e.g. "\\.rda"). Default is NULL.}
}
\value{
The dataframe with the summaries merged.
}
\description{
Merges the raw annotated coloc results into a summary dataframe.
}
