% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_eQTL_tidying.R
\name{tidy_eQTL_eQTLGen}
\alias{tidy_eQTL_eQTLGen}
\title{Tidy eQTLGen eQTLs for use with coloc.}
\usage{
tidy_eQTL_eQTLGen(eQTLGen_QTLs)
}
\arguments{
\item{eQTLGen_QTLs}{df. Full cis-eQTL summary statistics, as downloaded from
\url{https://www.eqtlgen.org/cis-eqtls.html}. Should be post 2019-12-23
changelog update.}
}
\value{
Tidy dataframe with eQTLGen eQTLs. Note that the dataframe includes
  an additional column (SNP_rs). This is necessary for MAF lookup, but should
  be removed prior to submitting to \code{colochelpR::get_coloc_results()}.
}
\description{
Tidy eQTLGen eQTLs for use with coloc.
}
