% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_general_functions.R
\name{save_coloc_results}
\alias{save_coloc_results}
\title{Save coloc results}
\usage{
save_coloc_results(coloc_results_annotated, results_dir_path)
}
\arguments{
\item{coloc_results_annotated}{Results, as returned by
\code{colochelpR::get_coloc_results}.}

\item{results_dir_path}{chr. The directory in which to save the merged
results.}
}
\value{
Saves coloc results in provided directory.
}
\description{
Save coloc results
}
